+++ src/dwarf_parser.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "types.h"
+
 #include <memory>
 #include <string>
 #include <vector>
@@ -11,6 +13,8 @@
 
 namespace bpftrace {
 
+class BPFtrace;
+
 class Dwarf
 {
 public:
@@ -21,18 +25,23 @@ public:
   std::vector<std::string> get_function_params(
       const std::string &function) const;
 
+  ProbeArgs resolve_args(const std::string &function);
+
 private:
   explicit Dwarf(const std::string &file_path);
 
-  std::unordered_map<std::string, Dwarf_Die> function_param_dies(
-      const std::string &function) const;
+  std::vector<Dwarf_Die> function_param_dies(const std::string &function) const;
 
   std::optional<Dwarf_Die> get_func_die(const std::string &function) const;
 
   std::string get_type_name(Dwarf_Die &type_die) const;
 
+  SizedType get_stype(Dwarf_Die &type_die) const;
+
   Dwfl *dwfl = nullptr;
   Dwfl_Callbacks callbacks;
+
+  std::string file_path_;
 };
 
 } // namespace bpftrace
@@ -46,7 +55,7 @@ namespace bpftrace {
 class Dwarf
 {
 public:
-  static std::unique_ptr<Dwarf> GetFromBinary(const std::string &file_path
+  static std::unique_ptr<Dwarf> GetFromBinary(const std::string &file_path_
                                               __attribute__((unused)))
   {
     static bool warned = false;
@@ -58,6 +67,11 @@ public:
 
   std::vector<std::string> get_function_params(const std::string &function
                                                __attribute__((unused))) const
+  {
+    return {};
+  }
+
+  ProbeArgs resolve_args(const std::string &function __attribute__((unused)))
   {
     return {};
   }